Restructure Board class
Board class now contains an 2D array of cells (pointers). It also has a 2D integer array to store filled values in a cell for each square, column, and row. For example, a filled value of 0b011000010 means that cells containing values 8, 7, and 2 exist within a cell.

Also wrote destructors for Cell and Board classes.